Fixes a bug in the paragraph ranges calculation.
